+++ src/App.svelte
@@ -2,15 +2,73 @@
     import SideBar from './layout/SideBar.svelte';
     import Content from './layout/Content.svelte';
     import GlobalNotifications from './layout/utils/notifications/GlobalNotifications.svelte';
+    import type { SidebarAction } from './layout/sidebar/sidebar-action';
+    import { SidebarActionOrientation } from './layout/sidebar/sidebar-action';
+    import { sidebarState } from './layout/stores/sidebar';
+    import { Settings } from './layout/settings';
+    import { fade } from 'svelte/transition';
+
+    let collapsedSidebar: boolean;
+    $: collapsedSidebar = $sidebarState === 'collapsed';
+
+    let settingsOpen: boolean = true;
+
+    const sidebarActions: SidebarAction[] = [
+        {
+            text: '&#9881;',
+            action: openSettings,
+            orientation: SidebarActionOrientation.Left
+        },
+        {
+            text: '&laquo;',
+            textCollapsed: '&raquo;',
+            action: toggleSidebar,
+            orientation: SidebarActionOrientation.Right
+        },
+    ];
+
+    /**
+     * Open settings view.
+     */
+    function openSettings(): void {
+        settingsOpen = true;
+    }
+
+    /**
+     * Close settings view.
+     */
+    function closeSettings(): void {
+        settingsOpen = false;
+    }
+
+    /**
+     * Toggle sidebar.
+     */
+    function toggleSidebar(): void {
+        sidebarState.set($sidebarState === 'normal' ? 'collapsed' : 'normal');
+    }
+
+    /**
+     * Refresh app.
+     */
+    function refreshApp(): void {
+        // @todo: Implement.
+    }
 </script>
 
 <div class="wrapper">
-    <div class="sidebar">
-        <SideBar/>
+    <div class="sidebar" class:collapsed={collapsedSidebar}>
+        <SideBar {sidebarActions} isCollapsed={collapsedSidebar}/>
     </div>
-    <div class="content">
-        <Content/>
+    <div class="content" class:wide={collapsedSidebar}>
+        <Content />
     </div>
+
+    {#if settingsOpen}
+        <div class="settings" transition:fade={{duration: 100}}>
+            <Settings on:closeSettings={closeSettings} on:dataUpdate={refreshApp}/>
+        </div>
+    {/if}
 
     <div class="global-notifications">
         <GlobalNotifications/>
@@ -22,17 +80,26 @@
 
     .wrapper {
         height: 100%;
-
         display: flex;
 
         .sidebar {
-            width: 20%;
+            flex-basis: 20%;
+            flex-grow: 2;
+
+            &.collapsed {
+                max-width: 3em;
+            }
         }
 
         .content {
             padding: 1em;
-            width: 80%;
+            flex-basis: 80%;
+            flex-grow: 8;
         }
+    }
+
+    :global(*) {
+        box-sizing: border-box;
     }
 
     :global(body) {
@@ -58,10 +125,25 @@
 
         &:hover {
             opacity: .75;
+            cursor: pointer;
+        }
+
+        &:active {
+            color: $button-text-color-active;
         }
     }
 
+    .settings {
+        position: fixed;
+        top: 0;
+        right: 0;
+        bottom: 0;
+        left: 0;
+        background: $color-background;
+    }
+
     .global-notifications {
+        z-index: 10;
         position: fixed;
         right: 1em;
     }