+++ src/layout/settings/index.ts
@@ -1,0 +1,5 @@
+import Settings from './Settings.svelte';
+
+export {
+    Settings
+}