+++ src/layout/sidebar/sidebar-action.ts
@@ -1,0 +1,11 @@
+export interface SidebarAction {
+    text: string;
+    action: () => void,
+    orientation: SidebarActionOrientation,
+    textCollapsed?: string;
+}
+
+export enum SidebarActionOrientation {
+    Left,
+    Right,
+}