+++ src/layout/utils/notifications/GlobalNotifications.svelte
@@ -32,6 +32,7 @@
 
         .notification {
             margin-bottom: 1em;
+            text-align: right;
         }
     }
 </style>