+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Changelog
 ---
 
+v1.3.0:
+--
+New features:
+- Global notifications for various actions;
+- Collapse sidebar to get that extra bit of space;
+- Settings view:
+  - App data export/import actions;
+  - App information data;
+
+Bugfixes:
+- Fixed "copy" functionality for sha hash;
+
+
 v1.2.0
 --
 New Apps:
@@ -11,10 +24,12 @@ Other:
 - Minor style adjustments for buttons
 - Global font change to monospace
 
+
 v1.1.0
 --
 - Updated app icon
 - Dark mode re-design
+
 
 v1.0.0
 --